requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
Pillow==10.1.0
pdf2image==1.16.3
rapidfuzz==3.8.0
python-dotenv==1.0.0
httpx==0.25.2 
fuzzywuzzy==0.18.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-Levenshtein==0.25.1 # For faster fuzzy matching

# Authentication
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-jose[cryptography]==3.3.0

# For testing
pytest
pytest-asyncio
fpdf